backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
ffmpeg-python>=0.2.0
pytesseract>=0.3.10
opencv-python-headless>=4.9.0
Pillow>=10.2.0
anthropic>=0.21.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
